Try-Catch Importing the Tagger
the Emscripten version of CPython doesn't support pickling, which is used for loading and saving the Unigram Tagger Pickle.

If we can't load dill at runtime, fallback to generating the tagger at runtime.

This isn't a solution for other problems, just for the module "multiprocessing", but will prevent it from crashing completely.